requirements.txt
Pillow==9.0.1
numpy==1.21.5
matplotlib==3.5.1
scipy==1.7.3